requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
cycler==0.10.0
DateTime==4.2
decorator==4.1.2
enum34==1.1.6
functools32==3.2.3.post2
ipdb==0.10.3
ipython==5.4.1
ipython-genutils==0.2.0
matplotlib==2.0.2
MySQL-python==1.2.5
numpy==1.13.1
olefile==0.44
pandas==0.20.3
pathlib2==2.3.0
pdb==0.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-gnupg==0.4.1
pytz==2017.2
PyYAML==3.12
scandir==1.5
scipy==0.19.1
seaborn==0.8
simplegeneric==0.8.1
six==1.10.0
subprocess32==3.2.7
traitlets==4.3.2
wcwidth==0.1.7
wordcloud==1.3.1
zope.interface==4.4.2
